+++ nautilus_trader/accounting/accounts/cash.pyx
@@ -347,7 +347,7 @@ cdef class CashAccount(Account):
             )
 
         if instrument.is_inverse and not use_quote_for_inverse:
-            return Money(commission, instrument.base_currency)
+            return Money(commission, instrument.get_base_currency())
         else:
             return Money(commission, instrument.quote_currency)